dashboard_backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2
pandas>=1.5.0
numpy>=1.23.0
requests>=2.28.0
apscheduler>=3.10.0
yfinance>=0.2.0
